重新生成user代码

+++ web/src/main/java/com/resms/ssm/auth/service/UserService.java
@@ -3,5 +3,5 @@ package com.resms.ssm.auth.service;
 import com.resms.ssm.auth.bean.User;
 
 public interface UserService {
-	public User getUserById(int id);
+	public User getUserById(long id);
 }
